feat: add HYPE token (#4156)
This is a temporary solution for getting the HYPE price. Currently,
HYPER is only on one source:
https://app.hyperliquid.xyz/trade/HYPE/USDC, and we decided to get the
price from the API. It will be used only for testnet. In the future, for
mainnet integration, we will build a solution based on Hyperliquid L1.
Added getting the price from the local cache if the fetching failed to
avoid triggering unnecessary alarms for the person on duty.