feat: add underpriced error message to underpriced errors (#4317)
When checking Sonic network I saw that rpc provider throws an error with
message `transaction underpriced` but relayer doesn't try to scale
transaction gas params but throws error `Unhandled error occurred during
iteration` and abandons transaction. In Sonic sometimes fetched
`rewardsPerBlockForPercentile` (it is always 1) is just not enough and
requires scaling up. It is not happening in current situation leading to
many transactions send by fallback or by none of the relayers resulting
in alarm