[RDS-1263] Send silent errors over critical-severity health check notifications (#4260)
Currently deployed changes work well, but are not covering the case when
the PagerDuty notification is triggered by health check alarms. These
are based on the lambdas throwing errors after their execution, so a
logic to only throw errors in case of high severity issues has been
implemented in following lambdas:
- lamba-ecs-services-monitoring
- lambda-hard-limits-monitoring
- lambda-lbtc-peg-checker
- lambda-monitoring-core
- lambda-monitoring-relayers
- lambda-node-price-deviation-monitoring

Silent errors will be sent to `SILENT_ERROR_NOTIFIER_LAMBDA_NAME`, which
for dev will be the same lambda as the `SEND_ERROR_LOG_LAMBDA_NAME` one,
but for prod it will go to a notifier sending PagerDuty notifications.

On top of that, monitoring changes need to be implemented, as Kuma will
send a pager duty notification with critical severity on monitoring
failure. Proposed approach - do not report a low-severity test failure
to Kuma, send silent error notification instead. Feel free to suggest a
different one if you come up with a better solution