[RDS-2140] Find out why monitoring for mqtt works on dev but not on prod (#4851)
The problem was that on prod the latency between when the package was
received to when the package was published was degrading over time
(getting higher and higher).

This problem was occuring because we were setting too many `setTimeouts`
in a short period of time. The number of `setTimeouts` set on prod was
around 2500 / sec, this number was lower on dev because dev only
listened for packages from 2 signers (prod from 5) - and that's why dev
was still able to work.

After refactoring the code so that the `setTimeout` is not set after
each new package is processed the problem disappeared.