Fix whitelist monitoring (#5195)
Image whitelist monitoring is currently not showing all of the missing
feeds, that's because the `requestDataPackages` function in the sdk
throws an error right when it sees first missing feed - instead of
aggregating the missing feeds and throwing the error with all of them.

So this PR changes this, it adds an `aggregateErrors` flag to the sdk
that makes the `requestDataPackages` function aggregate all of the
missing feeds first and only after that the errors is thrown.