feat: slippage is required again (#5669)
This solution might like look a little bit robust, however previous
solution was pretty fragile, because it was relying on the fact that:
empty array means that we are not calculating slippage (or missing
prices in local db), but undefined meant that there was error which
should not be ignored. Later this solution was (accidentally) removed in
#3813.

So I was aiming to build something hard to break, that is what I decided
to use this type: 'error' thing. However there is also another reason -
depth aggregator needs to treat different errors then slippage check, so
in my opinion best way is to propagate error to interested parties