feat: prepare SageOfChains to use in node only RPCs which are required (#5690)
This PR introduces changes to SageOfChains in preparation for the next
PR, where only the required RPCs defined in manifest sources will be
used. Currently, providers are built and exported in
blockchain-providers.ts, meaning their RPC URLs need to be defined when
the node is built. However, we determine which chains are required based
on the manifest defined in environment variables. This is why we need to
dynamically build and use the provider while validating that all
required RPCs are present in the NodeRunner constructor.
Key changes in this PR:
- SageOfChains is initialized once and cached.
- The SageOfChains constructor now creates a chainTypeAndId ->
providerFactory object, populated for every chain defined in the node
configuration.
- Switching from chainId to chainTypeAndId.
- The chainConfig is no longer stored in SageOfChains, as it can be
retrieved from the local chain configuration at any time.
- Providers are still built and exported in blockchain-providers, but
this will change in the following PR.