feat: use only required RPCs (#5696)
This PR adjust all fetchers and tests to dynamically build and use
providers only when it is required. This PR is closed coupled with
https://github.com/redstone-finance/redstone-monorepo-priv/pull/5690.

Copied from PR above
` Currently, providers are built and exported in
blockchain-providers.ts, meaning their RPC URLs need to be defined when
the node is built. However, we determine which chains are required based
on the manifest defined in environment variables. This is why we need to
dynamically build and use the provider while validating that all
required RPCs are present in the NodeRunner constructor.`