chore: imports unification (#6052)
I noticed that we use imports inconsistently. I decided to unify how we
import, providing some guidelines
- sometimes we use namespace imports (import * as) and sometimes default
imports (import x from) - when possible, it is better not to use any of
these forms at all but stick to named imports (import { a, b } from
'package') as they are more portable. I decided to prefer default
imports over namespace imports, but this is also an arbitrary decision.
- for nodejs-specific imports, we sometimes used 'node:' prefix and
sometimes not - I decided to go without the prefix, but the decision is
rather arbitrary
- I changed imports of our packages to use named imports always, but
changing everything to that form is a lot of work and probably not worth
it :), so not everything was changed
- a couple of times we had a library imported twice: with default import
and with named import. This was unified.