feat: dex depths lambda - add buy direction (#6081)
Problem:
Currently, lambda estimating price shift cost across different DEXes,
calculates it only for `sell` direction. It's because of the fact that
this is the only direction that is used to estimate market depth and
only these estimation values are saved in `marketDepthCache`. We need
both directions in the final script responsible for picking feeds which
are easy to manipulate.

Solution:
From now on, when depth aggregator is switched on - we're taking 6
slippage points for bot - sell and buy, creating two interpolations
based on these values and estimating depth based on the average from
these two calculations. Both of the interpolations are saved in the
market depth cache.