chore: remove ethereum-waffle (#6245)
Extract contracts creation to separate methods in fetchers and override
those methods in tests to use mock contracts.
Most of the time, we can get rid of the mock provider after such a
change.
In some tests, we call getBalance directly on the provider. In such
cases, we create a mock provider and add it to the sage of chains
object.