fix: curated list issue (#6377)
#### Fixes never-ending CuratedList timers.

Now, the timers start just after the first RPC is called, then the
quarantine timer starts after the first quarantine item is added
The timers stop working when no RPC is called since the last evaluation
period and when no item is quarantined.

It should'nt change drastically the behaviour when the RPCs are working
continuously, just the first item is released from the quarantine after
evaluation timeout + quarantine timeout.

Also, ClientBuilder is refactored to have all curated-list able to
enable and it's enabled only for Relayers (writable contract connectors)

---------

Co-authored-by: Jakub Jakubowski <7jakobjj@gmail.com>