chore: modify arweave worker lambda (#6423)
This modification solves problem with prod node sending data packages to
arweave lambda and receiving Too Long payload error. Its because we had
config of api gateway where we didn't need to wait lambda to respond but
we were returning 202 instantly. This type of config (async) has max
payload of 256 kb which we exceeded. So we are switching to "normal" api
gateway config (sync).

This solution allows the same - we still don't have to wait in node for
lambda completing update on arweave but required workaround with:
- node send request to api gateway -> api gateway redirects requests to
lambda -> lambda saves payload on S3 -> api gateway respond quickly with
OK
- another instance of lambda starts which is fetching data from s3 ->
saving to arweave - but that happens without on the background (for
node)