Contracts for perun (#6161)
A new adapter and price feed for Perun implementing the existing
interfaces.

In the adapter, we assume that:
- price updates can be made by 5 hardcoded updaters,
- updates may occur frequently, including multiple times within the same
block,
- we store price timestamps and block timestamps in microseconds,
- we support the concept of rounds; rounds are calculated with each
price update for a given feed,
- the price for a round is calculated according to the following
algorithm: we take the 3 most recent prices out of 5, calculate their
median, and if the newest price does not deviate from the median by more
than 1 percent, it is returned; otherwise, the median is returned.

The contract is optimized for gas efficiency. Alternatives I tested
include:
- other aggregation algorithms (the best one turned out to be the one
minimizing the number of comparisons),
- different methods of storing/reading data (using store/load in
assembly).

---------

Co-authored-by: Hatskier <Suvorovalex1997@gmail.com>