[RSFL-52] Split request to FastLane with multiple feeds into multiple requests with one feed (#7010)
There is a painful tx cost limit of 2M on hyperevm which is causing
problems with updating multiple feeds at once.
As a result we want to be able to send each feed in a seperate request
to FastLance auctions.
This will trigger concurrent auctions.
FastLance is expecting cost of a single tx to be no more than 70k on
hyperevm
Oev Relayer can now be configured using env field `OEV_MULTI_AUCTIONS`
to send feeds in a seperate calls to fastlance auctions.