feat: add fallback token mechanism and update uranium sources (#7223)
Added fallback token mechanism. Token can define which other price want
to use if fetching from all sources failed.

In that case it was added for URANIUM as tezos wants to us to use their
api and we want to have fallback to our previous sources. If tezos api
fails we use price from URANIUM_FALLBACK which uses different apis.