chore: unify hard limits monitoring code between monitoring and hard limits monitoring lambda (#7443)
We had two similar but different implementations of hard limits
monitoring: one in the monitoring package and one in the lambda
hard-limits-monitoring. This PR merges those two versions in a single
one that is now placed in monitoring. Lambda imports the check code from
there.

I tried to preserve all the features we have, e.g. in monitoring there
was a code that tried to refresh hard limits if some of them are not
available. This code runs only in ECS monitoring. On the other hand,
lambda reported problems with any of the hard limits files (we have two
copies) while ECS monitoring took one of the files and reported problems
with it. Now the lambda version is used in both environments.

There is one new thing: we check hard limits for tokens specified in the
fallback node manifest instead of all the tokens returned from the
gateway. This will allow a decrease in false alerts created by hard
limits during token introduction and decommission.