feat: adjust alltick api to fetch last price only when markets are open (#7704)
Alltick stocks api works in extended hours (before and after nasdaq and
nyse market) and we want to have data only when markets are operating in
standard hours. When markets are closed we want the last close price.

Introduced `IS_NY_STOCKS_MARKET_OPEN` which is taken from 3 different
sources, by voting from min 2 valid sources, and indicates if market is
open or not.