Remove build-dev script (#7504)
Remove `build-dev` script in favor of `typecheck`.

## Key points
 - `build-dev` used to output to `dist` directory
- `typecheck` **does not** output to dist directory, and simply runs
`tsc --noEmit` on all files as per `tsconfig.json` (usually including
tests and scripts)
- `build` works as previously (`tsc -p tsconfig.build.json`), i.e it
does typecheck and transpiles files to `dist` (as per
`tsconfig.build.json`)
 
## Implementation details
A new "transit node" in `turbo.json` is added: `setup-dev`. This is a
valid `turbo` task, but it doesn't exist in package.json. It simply
"depends" on building deps and compiling hardhat in current package.
Running `turbo setup-dev` in a package will prepare a dev environment
for you (but you can also `turbo build` if that's your thing - it
depends on `setup-dev` anyway)

## Benefits
- No more collisions between `build-dev` and `build` outputting to the
same `dist` - this should solve a whole class of bugs that we
encountered (but maybe not all of them - some are probably related to
remote caching).
- Performance: lighter dependency graph - extra dependencies on
`build-dev` previously caused almost double work in some cases

## Drawbacks
If someone is used to running `build-dev` to compile e.g. a script and
run `dist/script.js` - this is a breaking change for them.

Solution: use `ts-node`, `bun`, `deno`, `node
--experimental-strip-types`.

Note: this only applies to scripts and tests, you can still `build` an
app and run e.g `dist/index.js`

## Other changes in this PR
- Remove boilerplate `turbo.json` from all lambdas in favor of single
configuration in `build`
- Remove universal dependency on `^build` in `compile`. `compile` uses
`hardhat.config.ts`, which in some packages imports our libraries. This
means that you need these libs for hardhat config to be valid, **even
when it's not used for compile directly**. For now, I've hard-coded
these deps in a couple of packages, but the benefit is that everywhere
else we'll do much less re-`compile`s. This is a temporarily solution,
I'll think about better way to structure hardhat configs to mitigate
this in future.