feat(turbo): add lint:fix task (#7900)
https://redstone-team.atlassian.net/browse/RDS-4696

Ideally we'd like to have only `turbo lint` task and be able to pass the
args to it, like this:
```json
ROOT package.json
"lint:all:fix": "yarn turbo lint -- --fix"

/packages/*/package.json
"lint:fix": "yarn turbo lint -- --fix"
```
But this for some reason invalidates cache for other tasks like `build`
and `compile` (even though `--fix` is not passed to them, only to `lint`
task - I believe this is a bug). This still works, but it rebuilds
everything which is annoying, so for now we add explicit task for
`lint:fix`.