chore: fix oev multi auction (#7972)
fix oev multi auction: for now any error in auction (error is also when
there is no bid in auction) cause us not waiting for auction that was
successfull (we give fastlane few seconds to post tx). We should wait
even if 1 auction succeed. After all, when we have to post price - we do
it always after auctions anyways - so this code shouldn't break this.