feat: RDS-4942 add source timestamp metadata on price (#8069)
Till now, we have had a tentative property 'timestamp' on a returned
price from some sources. This was used for old price filtering (on
CEXes). Let's make it more permanent by moving it to 'metadata'. This
way source 'timestamp' will become available for monitoring and we will
be able to perform more sophisticated checks there.