feat: MultiFeedAdapter with reference (#7903)
Implementation of the MultiFeedAdapterWithoutRoundsWithReference
contract with tests. It implements `IMultiFeedAdapter` interface and
overrides `getLastUpdateDetails` implementation by querying two a main
adapter and a reference adapter as a sanity check.

- Designed to improve stability while introducing new, faster (but less
battle-tested) oracle nodes and relayers.
- Combines a new main adapter with a proven reference one: by default,
the fresher data is taken, expected to come from the main adapter. If
the main adapter fails or shows unexpected deviation, the reference
adapter is used instead.