Extract `getEcsEnv` function in `ecs-env-runner` to be reused in other package (#8653)
## Motivation
I intend to reuse `getEcsEnv` in Dry Run tests CI. For the Proof of
Concept in
https://github.com/redstone-finance/redstone-monorepo-priv/pull/8062 I
copied over and adapted code from `ecs-env-runner`. This PR will allow
to avoid committing duplicated code to `integration` package.

## Changes
* Made `runWithEcsEnv` non-interactive. It was prompting for user input
before and not really a "programatic API" as advertised in the
`README.md`
* Extracted and exported `getEcsEnv` function for fetching env vars and
secrets from ECS.

## Testing
* I run the `ecs-env-runner` CLI tool to fetch DEV Oracle Node envs on
`main` and on
https://github.com/redstone-finance/redstone-monorepo-priv/pull/8653/commits/b04cb46d30c35ba8044b58e0e89fb04c063cfd10
and got exactly the same output.