feat: updated Adapter With Reference to support MegaETH (#8738)
Added support for timestamps in different formats to support MegaETH
integration, where:
- main adapter (FastMultiFeedAdapter) will return
  - data timestamp in microseconds
  - block timestamp in microseconds
- ref adapter (MultiFeedAdapter) will return
  - block timestamps in seconds
  - data timestamps in milliseconds

Added corresponding tests. Default implementation of
`convertDataTimestampFromAdapter` and `convertBlockTimestampFromAdapter`
ensure that previous logic (used on the Monad network) remains
unchanged. Old tests also passed.

On monad, for comparison, all block timestamps on both adapters are in
seconds, and data timstamps are in microseconds


[EDIT] After long discussions with @szynwelski we decided to minimise
the changes on the AdapterWithReference and create a simple proxy for
the ref adapter that will be responsible for the timestamps unification.
We will then just assume that AdapterWithReference can work correctly
only with consistent adapters.