Run "Long dry run test" with the same env vars as Oracle Nodes on DEV (#8735)
## Motivation
"Long dry run test" spawns an Oracle Node as a subprocess. The node
started in the test is run with a minimal config. The results produced
by the node are later compared with the results produced by a node
running in production. There are a couple of problems with this
approach:
* The node run in tests uses a fixed hardcoded configuration that can be
vastly different from the current configuration of the production node.
* The node run in tests relies on a set of default, hardcoded, usually
free and public blockchain RPCs and source data APIs, as opposed to
production node that relies on paid 3rd party dependencies. As a result,
the data produced in test will often differ from the prod data. The
instability of the "free" 3rd party dependencies causes undeterministic
failures of the test.

This PR changes the configuration of the Oracle Node started in "Long
dry run test" to mirror the configuration of a node running in DEV
environment. The goal is to solve the two problems mentioned above.

## Changes
* Fetch env vars and secret vars that the DEV Oracle Node is started
with from AWS ECS
* Add logic for keeping track of fetched variables. If configuration of
DEV Oracle Node changes it will be easy to adjust "Long dry run test"
configuration.
* Print the results of ECS variables resolution to GH Actions job
summary and logs.
* Remove passing of Oracle Node variables from GH Workflow file. All
Oracle Node env vars are now resolved in a single place.
* Lower `MAX_ACCEPTED_DEVIATION_PERCENT` to `2`. 
* Increase `acceptablePerTokenFailuresPercent` to `20` for
`primary.json` manifest.
* Remove all (but TWAP) tokens from `DATA_FEEDS_NOT_WORKING_LOCALLY`.
Re-add tokens to skip based on test results on new config.

## Example test run
Commit
https://github.com/redstone-finance/redstone-monorepo-priv/pull/8735/commits/410452503e6c05a9cda400897fed8b513cf79e8e:
https://github.com/redstone-finance/redstone-monorepo-priv/actions/runs/20069453273

## Possible future improvements
* Limit the number of possible job reruns
* The test runs on "paid" 3rd party dependencies. We can limit the
number of times the CI job is "rerun".
* Finish test early if it is going to fail
    * Same motivation as above - to potentially limit costs.
* Check for missing sources in metadata (not only missing data packages)